# SaveState: Copy RAM using threads