On my (powerful) desktop, this reduces SaveStart::DoState from 17.5ms to 7.5ms.  This is significant for rewind, because it takes it from "slower than an entire frame" to "might fit within a frame".

Doing VRAM as well went from about 9.3ms -> 8.5ms, and doing SIMD took it from 8.5ms -> 7.5ms.  With that, it's actually taking about 50% time to do all the other non-memory stuff, so there could be other places to optimize.

Removing the SIMD copy and switching to memmove is slower, but still better than before.  My averages are all trending lower now (maybe because of zstd?, I'm measuring outside it but repeating the test which includes it), but but it's around 3-4ms slower than SIMD in the common case, which is still good.

This is faster on mobile too (around 32ms -> 23ms on a Pixel.)

-[Unknown]