@hrydgard 2021-04-13
Those are some weird speed differences! But certainly nicer code this way.